requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
beautifulsoup4==4.9.1
black==19.10b0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
pathspec==0.8.0
pylint==2.5.3
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
